+++ packages/apptive_grid_core/lib/model/data_entity.dart
@@ -269,3 +269,23 @@ class AttachmentDataEntity extends DataEntity<List<Attachment>, dynamic> {
   @override
   int get hashCode => toString().hashCode;
 }
+
+/// [DataEntity] representing [Geolocation]s
+class GeolocationDataEntity extends DataEntity<Geolocation, dynamic> {
+  /// Creates a new GeolocationDataEntity Object with value [value]
+  GeolocationDataEntity([Geolocation? value]) : super(value);
+
+  /// Creates a new GeolocationDataEntity Object from json
+  /// [json] needs to be an array of double
+  factory GeolocationDataEntity.fromJson(dynamic json) {
+    Geolocation? jsonValue;
+    if (json != null) {
+      jsonValue = Geolocation.fromJson(json);
+    }
+    return GeolocationDataEntity(jsonValue);
+  }
+
+  /// Returns [value] coordinates in a array
+  @override
+  dynamic get schemaValue => value?.toJson();
+}