+++ packages/apptive_grid_form/lib/widgets/form_widget/form_widget.dart
@@ -45,6 +45,10 @@ Widget fromModel(FormComponent component) {
       return EnumCollectionFormWidget(
         component: component as EnumCollectionFormComponent,
       );
+    case GeolocationDataEntity:
+      return GeolocationFormWidget(
+        component: component as GeolocationFormComponent,
+      );
     default:
       throw ArgumentError(
         'No Widget found for component $component. Please make sure you are you using the latest version of this package?',